Moving on to BHP
For the most part I have already incorporated many of the most useful lessons from Violent Python into the Arms Commander Project. Everything else in the book (Violent Python) seems it requires a lot of hands-on research to make it effective. For that reason, we are moving on to lessons to be learned from Black Hat Python. Note a lot of the source code must be modified to do what the book claims it should do. My home/user is clearly not Justin as my Kali Linux username in "pic_carver.py" for example. You will have to do the same thing, change things up, customize the script and see what it does.